Move getOrCreateStream to its own file
